src/requirements.txt
Django==1.10.3
django-crispy-forms==1.6.1
psycopg2==2.6.2
django-datetime-widget==0.9.3
django-tables2==1.4.2